Add lint workflow using jshint@2.12.0
Add .jshintignore to skip slow/irrelevant linting of node_modules and
vendor js.

Add jshint@2.12.0 as a top-level devDependency to guarantee it's
available for use from the shell. Upgrade grunt-contrib-jshint to use
jshint@2.12.0 and pin grunt at ^1.0.3 since that's already in the
lockfile and is required by grunt-contrib-jshint@3.0.0.

Update grunt config to watch/jshint the same files as standalone jshint